Fixed some issues with SqlJobQueue.

+++ src/com/staticbloc/jobs/SqlJobQueue.java
@@ -10,12 +10,14 @@ import android.util.Log;
 import com.google.gson.Gson;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 public final class SqlJobQueue extends BasicPersistentJobQueue {
     private final DbHelper db;
 
-    SqlJobQueue(Context context, JobQueueInitializer initializer) {
+    public SqlJobQueue(Context context, JobQueueInitializer initializer) {
         super(context, initializer);
         db = DbHelper.getInstance(context.getApplicationContext(), "com.staticbloc.jobs.db");
     }
@@ -45,7 +47,7 @@ public final class SqlJobQueue extends BasicPersistentJobQueue {
         getPersistenceExecutor().execute(new Runnable() {
             @Override
             public void run() {
-                db.removeJob(getName(), job);
+                db.removeJob(job);
             }
         });
     }
@@ -65,12 +67,14 @@ public final class SqlJobQueue extends BasicPersistentJobQueue {
         getPersistenceExecutor().execute(new Runnable() {
             @Override
             public void run() {
-                db.updateJob(getName(), job);
+                db.updateJob(job);
             }
         });
     }
 
     private static class DbHelper extends SQLiteOpenHelper {
+        private static final Map<JobQueueItem, Long> jobToSqlIdMap = new HashMap<>();
+
         private static volatile DbHelper instance = null;
 
         public static DbHelper getInstance(Context context, String name) {
@@ -89,14 +93,12 @@ public final class SqlJobQueue extends BasicPersistentJobQueue {
         private static final String KEY_QUEUE_NAME = "queue_name";
 
         private static final String JOB_TABLE_NAME = "jobs";
-        private static final String KEY_JOB_ID = "job_id";
         private static final String KEY_JOB_JSON = "job_json";
         private static final String CREATE_JOB_TABLE = "CREATE TABLE IF NOT EXISTS " + JOB_TABLE_NAME + " (" +
                 KEY_ROW_ID + " INTEGER PRIMARY KEY AUTOINCREMENT, " +
                 KEY_QUEUE_NAME + " TEXT NOT NULL, " +
-                KEY_JOB_ID + " INTEGER NOT NULL, " +
                 KEY_JOB_JSON + " TEXT NOT NULL);";
-        private static final String JOB_WHERE_CLAUSE = KEY_QUEUE_NAME + "='%s' AND " + KEY_JOB_ID + "=%d";
+        private static final String JOB_WHERE_CLAUSE = KEY_ROW_ID + "=%d";
         private static final String ALL_JOBS_IN_QUEUE_WHERE_CLAUSE = KEY_QUEUE_NAME + "='%s'";
 
         private Gson gson;
@@ -113,7 +115,7 @@ public final class SqlJobQueue extends BasicPersistentJobQueue {
             Cursor c = null;
             try {
                 c = getWritableDatabase().query(JOB_TABLE_NAME,
-                        new String[] {KEY_JOB_JSON},
+                        new String[] {KEY_ROW_ID, KEY_JOB_JSON},
                         String.format(ALL_JOBS_IN_QUEUE_WHERE_CLAUSE, queueName),
                         null,
                         null,
@@ -122,7 +124,9 @@ public final class SqlJobQueue extends BasicPersistentJobQueue {
                 if(c.moveToFirst() && !c.isLast()) {
                     List<JobQueueItem> jobs = new ArrayList<>(c.getCount());
                     do {
-                        jobs.add(gson.fromJson(c.getString(c.getColumnIndex(KEY_JOB_JSON)), JobQueueItem.class));
+                        JobQueueItem job = gson.fromJson(c.getString(c.getColumnIndex(KEY_JOB_JSON)), JobQueueItem.class);
+                        jobs.add(job);
+                        jobToSqlIdMap.put(job, c.getLong(c.getColumnIndex(KEY_ROW_ID)));
                     } while(c.moveToNext());
                     return jobs;
                 }
@@ -146,26 +150,35 @@ public final class SqlJobQueue extends BasicPersistentJobQueue {
             queueName = DatabaseUtils.sqlEscapeString(queueName);
             ContentValues values = new ContentValues();
             values.put(KEY_QUEUE_NAME, queueName);
-            values.put(KEY_JOB_ID, job.getJobId());
             values.put(KEY_JOB_JSON, DatabaseUtils.sqlEscapeString(gson.toJson(job)));
-            getWritableDatabase().insert(JOB_TABLE_NAME, null, values);
+            long id = getWritableDatabase().insert(JOB_TABLE_NAME, null, values);
+            jobToSqlIdMap.put(job, id);
         }
 
-        public void updateJob(String queueName, JobQueueItem job) {
-            queueName = DatabaseUtils.sqlEscapeString(queueName);
+        public void updateJob(JobQueueItem job) {
+            Long id = jobToSqlIdMap.get(job);
+            if(id == null) {
+                // TODO: log warning
+                return;
+            }
             ContentValues values = new ContentValues();
             values.put(KEY_JOB_JSON, DatabaseUtils.sqlEscapeString(gson.toJson(job)));
             getWritableDatabase().update(JOB_TABLE_NAME,
                     values,
-                    String.format(JOB_WHERE_CLAUSE, queueName, job.getJobId()),
+                    String.format(JOB_WHERE_CLAUSE, id),
                     null);
         }
 
-        public void removeJob(String queueName, JobQueueItem job) {
-            queueName = DatabaseUtils.sqlEscapeString(queueName);
+        public void removeJob(JobQueueItem job) {
+            Long id = jobToSqlIdMap.get(job);
+            if(id == null) {
+                // TODO: log warning
+                return;
+            }
             getWritableDatabase().delete(JOB_TABLE_NAME,
-                    String.format(JOB_WHERE_CLAUSE, queueName, job.getJobId()),
+                    String.format(JOB_WHERE_CLAUSE, id),
                     null);
+            jobToSqlIdMap.remove(job);
         }
 
         public void removeAllJobs(String queueName) {
